Refactor fields location
